+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 中文 | [英文](https://github.com/FightingDesign/fighting-design/blob/master/CHANGELOG.en-US.md)
 
+**优化**
+
+- 优化 `f-trigger` 和 `f-dropdown` 样式细节
+
 ## 0.35.0 (2023-05-07)
 
 **优化**